requirements.txt
python-dotenv~=0.19.1
Flask~=2.0.1
gensim~=4.1.2